Don't save skipped tasks in the job log; move the delegate configurators conditional processing logic.
Partially revert the previous commit. Because most of the time we want to compare the current operation's digest with the last one that actually modified the resource. Instead allow the delagate configurator to skip a subtask when it runs so its digest will be saved in the job log even while skipping the actuall operation.