loader, import: don't add environment instances or imports to ensemble except when exporting environments.
By default, don't environment instances unless they are imported from external ensembles.
Similarly, don't include imports declared in an environment unless they have a namespace_prefix that reference by a connection in the environment.
This can be overridden by setting the 'load_env_instances' in LocalEnv.overrides, which is when environments export does to include the instances in the export.