server: track commit dates in cache for shallow clones
This fixes an issue where we return stale content because we can't determine if a cached commit has changed.
This could happen if the (shallow) local clone is recreated after a value is cached.
By remembering the commit date in the cache we can assure that the clone is shallow since that date.