feat(plan): add "customized" to instance status to indicate that it has intentionally diverged from its template and the reconfigure check should be skipped.
It is automatically set when a custom operation (including "discover" operations) modifies an attribute that was either set by or a dependency of the previous configure operation. This prevents a future configure operation from overwriting those changes.

Setting the `--change-detection` job option to "always" will ignore the "customized" field and potentially the configure operation will re-run and overwrite custom changes to the instance

Add a "digestPut" key to job log to remember attributes that were set by a task.